# healthai-API/MSPR1-API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MSPR1 API",
        "description": "API for exercise, food, and meal tracking",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "MSPR1 API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/exercises": {
            "delete": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Supprimer un exercice",
                "operationId": "b53ced622476faaef89a54475c7c7fd6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resources"
                                ],
                                "properties": {
                                    "resources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercise deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/foods": {
            "delete": {
                "tags": [
                    "Foods"
                ],
                "summary": "Supprimer un aliment",
                "operationId": "4adcd0ed6478723690b678b3868cd479",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resources"
                                ],
                                "properties": {
                                    "resources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Food deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/sessions": {
            "delete": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Supprimer une session",
                "operationId": "cd86d94d3828f2c21fddd9ac95ece896",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resources"
                                ],
                                "properties": {
                                    "resources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}